requirements.txt
pytelegrambotapi
pillow
englisttohindi
flask
threading
